UVa 612 - DNA Sorting
You are responsible for cataloguing a sequence of DNA strings (sequences
containing only the four
letters A, C, G, and T). However, you want to catalog them, not in
alphabetical order, but rather in
order of “sortedness”, from “most sorted” to “least sorted”. All the
strings are of the same length.