UVa 272 - TEX Quotes
Input will consist of several lines of text containing an even number of
double-quote (") characters. Input
is ended with an end-of-file character. The text must be output exactly
as it was input except that: the first " in each pair is replaced by two
` characters: `` and the second " in each pair is replaced by two '
characters: ''.